fix(emilib): fix _mlf overflow causing rehash failure and CI hang
In emilib2/3, using size_t = uint32_t caused uint32_t * uint32_t overflow in required_buckets calculation, preventing rehash from triggering and causing find_empty_slot infinite loop. Also add timeout protection to CI test steps.

+++ include/emilib/emihmap3.hpp
@@ -589,7 +589,7 @@ public:
     }
 
     template <typename K, typename V> size_t insert_unique(K&& key, V&& val) noexcept {
-        const size_t required_buckets = ((size_t)_num_filled * _mlf >> 28);
+        const size_t required_buckets = ((uint64_t)_num_filled * _mlf >> 28);
         if (required_buckets >= _num_buckets)
             rehash(required_buckets + 2);
 
@@ -915,7 +915,7 @@ private:
     // Find the bucket with this key, or return a good empty bucket to place the key in.
     // In the later case, the bucket is expected to be filled.
     template <typename K> size_t find_or_allocate(const K& key, bool& bnew) noexcept {
-        const size_t required_buckets = ((size_t)_num_filled * _mlf >> 28);
+        const size_t required_buckets = ((uint64_t)_num_filled * _mlf >> 28);
         if (required_buckets >= _num_buckets)
             rehash(required_buckets + 2);
